fixed some warnings

+++ src/network_packet.c
@@ -90,10 +90,10 @@ av_packet* to_av_packet(HTTP_packet* network_packet)
 	memcpy(av, network_packet->message + 1, sizeof(av_packet));
 	av->av_data.data = malloc(av->av_data.size);
 	uint8_t * data = malloc(av->av_data.size);
-	printf("av->av_data.size %d\n", av->av_data.size);
-	printf("av->av_data.data 0x%x\n", av->av_data.data);
+	printf("av->av_data.size %d\n", (unsigned int) av->av_data.size);
+	printf("av->av_data.data 0x%x\n", (unsigned int) av->av_data.data);
 	av->av_data.data = data;
-	printf("(malloc'd) av->av_data.data 0x%x\n", av->av_data.data);
+	printf("(malloc'd) av->av_data.data 0x%x\n",(unsigned int)av->av_data.data);
 	memcpy(av->av_data.data, network_packet->message + 1 + sizeof(av_packet), av->av_data.size);
 	return av;
 }
